Written by: Todd Fries <todd@openbsd.org>
OpenCM alpha16

OpenCM is designed as a secure, high-integrity replacement for CVS. A list of
the key features can be found on the features page. While not as ``feature
rich'' as CVS, it supports some useful things that CVS lacks. Briefly, OpenCM
provides first-class support for renames and configuration, cryptographic
authentication and access control, and first-class branches. 

The OpenCM project was originally started because we needed a secure,
high-integrity configuration management system for the EROS project.
Alternatives, such as BitKeeper, Subversion, and PerForce, either did not
meet our requirements or were not available at the time the work started. We
had previously used CVS, but it's absence of real branches and configurations
finally drove us to build a better tool.