Written by: Nikolay Sturm <sturm@openbsd.org>
import of unison 2.9.1
submitter Zvezdan Petkovic

Unison is a file-synchronization tool for Unix and Windows.  It allows
two replicas of a collection of files and directories to be stored on
different hosts (or different disks on the same host), modified
separately, and then brought up to date by propagating the changes in
each replica to the other.

naddy@ OK