Written by: Jakob Schlyter <jakob@openbsd.org>
pflogsumm is a log analyzer/summarizer for the Postfix MTA.  It is
designed to provide an over-view of Postfix activity, with just enough
detail to give the administrator a "heads up" for potential trouble spots.

pflogsumm generates summaries and, in some cases, detailed reports of mail
server traffic volumes, rejected and bounced email, and server warnings,
errors and panics.