Written by: Brad Smith <brad@openbsd.org>
Move headers path back to the intended spot, it was just too easy
to have them picked up in the current spot instead of Gtk+2 headers.
This was originally done to prevent potential breakage from an upgrade
when the headers path was changed to prevent this kind of problem in
the first place but it looks like all ports deal with this gracefully
since they use "gdk-pixbuf-config --cflags" to find the path.