Written by: Anil Madhavapeddy <avsm@openbsd.org>
initial import of check-0.8.4
--

Check is a unit test framework for C. It features a simple interface
for defining unit tests, putting little in the way of the developer.
Tests are run in a separate address space, so Check can catch both
assertion failures and code errors that cause segmentation faults
or other signals. The output from unit tests can be used within
source code editors and IDEs.