Written by: Todd Fries <todd@openbsd.org>
Import eel-2.0.8
--
The Eazel Extensions Library is a collection of widgets and extensions
to many modules of the GNOME platform. These widgets and extensions
were developed by hackers working on Nautilus. For the duration of
the Nautilus 1.0 development cycle, the code was internal to Nautilus
and its components.

From marcm for gnome2