Written by: Margarida Sequeira <margarida@openbsd.org>
Import of neXtaw-0.13.0
--
neXtaw is a replacement library for the Athena (libXaw) widget set.
It is based on Xaw3d, by Kaleb Keithley and is almost 100% backward
compatible with it. Its goal is to try to emulate the look and feel
of the N*XTSTEP GUI.

WWW: http://siag.nu/neXtaw/

brad@ ok