Written by: Dan Harnett <danh@openbsd.org>
import of cronolog 1.6.2
--
cronolog is a simple filter program that reads log file entries from
standard input and writes each entry to the output file specified by a
filename template and the current date and time.  When the expanded
filename changes, the current file is closed and a new one opened.
cronolog is intended to be used in conjunction with a web server, such
as Apache, to split the access log into daily or monthly logs.

WWW: http://www.cronolog.org