Written by: Margarida Sequeira <margarida@openbsd.org>
Import of sleuthkit-1.62
--
The Sleuth Kit (previously known as TASK) is a collection of 
UNIX-based command line file system forensic tools that allow 
an investigator to examine NTFS, FAT, FFS, EXT2FS, and EXT3FS 
file systems of a suspect computer in a non-intrusive fashion. 
The tools have a layer-based design and can extract data from 
internal file system structures. Because the tools do not rely 
on the operating system to process the file systems, deleted 
and hidden content is shown.

This port replaces TASK, previously removed.