Written by: Margarida Sequeira <margarida@openbsd.org>
Initail import of ipaudit-0.95
--
IPaudit is a software package to record and display network activity. 
It includes ipaudit, which stores counts of bytes and packets for 
every combination of host/port pairs and protocol. 
The utilities total and ipstrings can be used to investigate network 
traffic records from the command line.

WWW: http://ipaudit.sourceforge.net/