Written by: Christian Weisgerber <naddy@openbsd.org>
Import mod_random 1.4.
Submitted by Xavier Santolaria <xavier@santolaria.net>.

mod_random provides three services.  The first service is as a
redirector.  The second is that it provides environmental variables
that can be used for doing ad banner systems.  The third is that it
can be used to display entire pages of random html.