Written by: Anil Madhavapeddy <avsm@openbsd.org>
initial import of p5-Pod-Simple-0.96
--
Pod::Simple is a Perl library for parsing text in the Pod ("plain
old documentation") markup language that is typically used for
writing documentation for Perl and for Perl modules. The Pod format
is explained in the perlpod man page; the most common formatter is
called "perldoc".