Written by: Margarida Sequeira <margarida@openbsd.org>
Security Fix:
A problem with Ghostscript could make it possible to execute
arbitrary commands. The vulnerability exists when GhostScript
is used to process specially formatted PS files.
An attacker can exploit this vulnerability by creating a
malicious PS file which, when processed with GhostScript, will
result in the execution of arbitrary system commands.

Bump version to 7.05p1

ok naddy@