Written by: Anil Madhavapeddy <avsm@openbsd.org>
initial import of autobook-1.3 ; this is based on the 1.2 port that used
to live in devel/autobook.  espie@ ok the move, no response on the
upgrade to v1.3

--
This book is a tutorial for Autoconf, Automake and Libtool, hereafter
referred to as the GNU Autotools. The GNU manuals that accompany
each tools adequately document each tool in isolation. Until now,
there has not been a guide that has described how these tools work
together.