Written by: Brad Smith <brad@openbsd.org>
upgrade to binutils 2.13.2.1

This is work in progress. It is being commited as a starting point so other
people can look at it/fix issues/and maybe Art can add Hammer support? :)
It has been lightly tested on macppc/sparc64/i386 and compile tested by
mickey@ on hppa.

DO NOT try to build ld.so on sparc/sparc64/macppc and install the resulting
binary unless you like having a hosed system. A known "misfeature" in ld.so
that is being worked on.
--
MAINTAINER ok