Written by: Marc Matteo <marcm@openbsd.org>
Update to 0.3.12 and regen patches.

In addition to the usual bug fixing this update does away with the hack
that built a custom XML catalog file (now that we have proper XML catalog
support).  This will cause some problems with scrollkeeper dependent ports
until I commit the fix for them shortly.