Written by: Jolan Luff <jolan@openbsd.org>
import of stress 0.18.1:

stress is a tool which imposes a configurable amount of CPU, memory,
I/O, or disk stress on a POSIX-compliant operating system.  It is
written in highly-portable ANSI C, and uses the GNU Autotools to
compile on a great number of UNIX-like operating systems.  stress is not
a benchmark.  It is a tool used by system administrators to evaluate how
well their systems will scale, by kernel programmers to evaluate
perceived performance characteristics, and by systems programmers to
expose the classes of bugs which only or more frequently manifest
themselves when the system is under heavy load.