Written by: Margarida Sequeira <margarida@openbsd.org>
Security fix:
Two bugs were discovered that lead to a buffer overflow in PostgreSQL
in the abstract data type (ADT) to ASCII conversion functions.
It is believed that, under the right circumstances, an attacker may use
this vulnerability to execute arbitrary instructions on the PostgreSQL
server.

ok Brandon Palmer (maintainer)