Written by: Jolan Luff <jolan@openbsd.org>
on macppc, when the battery is fully charged and the machine is on a/c
power, the time remaining to charge alternates between 0 minutes and 2+
months.  this makes the panel grow and shrink often (to display 2 or 4
digits) and is quite distracting.  so if we get a large charge time,
just zero the buffer.