Written by: Brad Smith <brad@openbsd.org>
Oneliner to add use of msync().

"On OpenBSD, data read from a mmap()ed file through read() is not in sync
unless msync() is explicitely called.

In pure-ftpd it breaks the -y option that restricts the number of sessions
an account can open. This oneliner seems to fix it."

ok pvalchev@
--
From: Frank Denis <j at pureftpd dot org>