Written by: Nikolay Sturm <sturm@openbsd.org>
Initial import of onesixtyone 0.3.2

onesixtyone takes a different approach to SNMP scanning. It takes advantage
of the fact that SNMP is a connectionless protocol and sends all SNMP
requests as fast as it can. Then the scanner waits for responses to come
back and logs them, in a fashion similar to Nmap ping sweeps.

from Srebrenko Sehic <haver at insecure.dk>