Written by: Xavier Santolaria <xsa@openbsd.org>
Import of py-cheetah-0.9.15

Cheetah is a Python-powered template engine and code generator. It can
be used as a standalone utility or it can be combined with other tools.
Cheetah has many potential uses, but web developers looking for a viable
alternative to ASP, JSP, PHP and PSP are expected to be its principle
user group.

ok sturm@.