Written by: Mathieu Sauve-Frankel <msf@openbsd.org>
Initial import of ExtUtils-CBuilder-0.02

This module can build the C portions of Perl modules by invoking the
appropriate compilers and linkers in a cross-platform manner.  It was
motivated by the "Module::Build" project, but may be useful for other
purposes as well.  However, it is not intended as a general
cross-platform interface to all your C building needs.

ok sturm@