Written by: Anil Madhavapeddy <avsm@openbsd.org>
initial import of ocaml-postgresql-1.3.3
--
This OCaml library provides an interface to PostgreSQL (tested with
versions 7.0.2, 7.1.3 and 7.4.1), an efficient and reliable, open
source, relational database.

Almost all functionality available through the C-API (libpq) is
replicated in a type-safe way. This library uses objects for
representing database connections and results of queries.