Written by: Peter Galbavy <peter@openbsd.org>
Update to PostgreSQL 7.4.3

Changes in the package layout means that there is now a -client and
-server subpackage, and users no longer need to decide which, as
the -server depends on the -client, just like the sane cases in
mysql and openldap.

Other changes include:

* Removal of the tcl FLAVOR until someone with more tcl/tk knowledge
can make it work correctly.

* The INSTALL-server script now created a _postgresql user and group
if they don't already exist, and also a default database in
/var/postgresql if that directory also doesn't exist.

* The port is marked for NO_SHARED_ARCHS as the -server subpackage
needs shared lib support to build.

* The port will build with spinlocks disabled on hppa until someone
can check and test this problem more closely.

Built and checked on i386, sparc64, amd64, macppc (waiting for regress
test feedback).

Dependent packages will be updated after this commit, shortly.