Written by: Marc Matteo <marcm@openbsd.org>
Update to 2.6.0.  Get rid of the custom OpenBSD support from the
former nino@ that has been a part of this port for some time.
Updates to the internal API finally broke it, but a huge thanks
to Nils for his work there (thanks Nils!).

Now we use the Free/NetBSD support as a base and patch from there :).