Written by: Marc Matteo <marcm@openbsd.org>
Update to 0.10.11.  Yep, a commit as long as your arm for a 0.00.01
version number bump.  Yeah, well, it took me a bit to even get it
to compile, one of those, "Hey, let's change the entire build
infrastructure for a minor bugfix!" things...

But hey now we're fully in-sync with Gnome 2.6.2.