Written by: Brad Smith <brad@openbsd.org>
This is ridiculous, I remove this port after no updates for 2 yrs because
the version in the ports tree had security issues. Then a couple weeks
later it is revived and then 2 security issues pop up just before release.
This pile of shit is being marked BROKEN and frankly I don't care if it
ever comes back.

ok pvalchev@