Written by: Xavier Santolaria <xsa@openbsd.org>
Initial import of py-feedparser-3.3.

Universal Feed Parser is a Python module for downloading and parsing
syndicated feeds. It can handle RSS 0.90, Netscape RSS 0.91, Userland
RSS 0.91, RSS 0.92, RSS 0.93, RSS 0.94, RSS 1.0, RSS 2.0, Atom, and CDF
feeds.

Universal Feed Parser is easy to use; the module is self-contained in a
single file, feedparser.py, and it has only one public function, parse.
parse takes a number of arguments, but only one is required, and it can
be a URL, a local filename, or a raw string containing feed data in any
format.