Written by: Nikolay Sturm <sturm@openbsd.org>
Under unclear circumstances parts of tk are recompiled in fake stage.
In this case using PREFIX in a Makefile to find installed files is
obviously wrong. In 8.3 and 8.4 this part of the patch is not necessary
at all, so just remove it. In 8.0 use TCL_LIB_SPEC, as is used in 8.3
and 8.4.

problem noted by Rod.. Whitworth <listen at witworx.com>