Written by: Mathieu Sauve-Frankel <msf@openbsd.org>
Initial import of Locale-Maketext-Gettext

Locale::Maketext::Gettext joins the GNU gettext and Maketext frameworks.
It is a subclass of Locale::Maketext(3) that follows the way GNU gettext
works. It works seamlessly, both in the sense of GNU gettext and
Maketext. As a result, you enjoy both their advantages, and get rid of
both their problems, too.