Written by: Kevin Lo <kevlo@openbsd.org>
Initial import of p5-VCP-autrijus 0.9

vcp (and the Perl module VCP.pm) is a system for copying resources under
version control between repositories. CVS, p4 and VSS repositories are
currently supported.
Note that this is not a branch, nor a fork, from VCP's main repository
This is merely a snapshot of versions that are known to work with svk.