Written by: Brad Smith <brad@openbsd.org>
rev 1.28

Propagate a convenience library's dependency libs correctly when
it is being linked into a libtool library.  Deplibs are now
propagated whether libdir is set or not.

rev 1.27

The dlpreopen pass over libraries reverses the elements in the
$deplibs list.  This causes problems when the link pass tries to
find libraries when they are located in non-standard places
denoted by -L options.  Due to the reversed order these -L options
occur after the libraries that need them, and they are not found:

(Un)Reverse $deplibs list at the start of the link pass in lib mode.

From libtool CVS