Written by: Nikolay Sturm <sturm@openbsd.org>
Initial import of p5-XML-RSS-Parser 2.15

XML::RSS::Parser is a lightweight liberal parser of RSS
feeds. This parser is "liberal" in that it does not
demand compliance of a specific RSS version and will
attempt to gracefully handle tags it does not expect or
understand.

from Sam Smith <s at msmith.net>