Written by: Nikolay Sturm <sturm@openbsd.org>
Initial import of Params-CallbackRequest 1.13

Params::CallbackRequest provides functional and object-oriented callbacks
to method and function parameters. Callbacks may be either code references
provided to the new() constructor, or methods defined in subclasses of
Params::Callback. Callbacks are triggered either for every call to the
Params::CallbackRequest execute() method, or by specially named keys in
the parameters to execute().